refactor: popular movie api 함수 분리
