+++ src/api/getPopularMovieResult.ts
@@ -1,0 +1,7 @@
+import baseApi from "./baseApi";
+
+export const getPopularMovieResult = async (page: number) => {
+  const movieResult = await baseApi("/movie/popular", page);
+
+  return movieResult;
+};
